refactor: move send-event to event send, deprecate send-envelope
- Canonical command is now 'sentry event send' under the event route
- send-envelope is now a deprecation shim suggesting 'sentry event send --raw'
- send-event remains as a hidden backward-compat alias
- Removed the 'send' route group entirely
- Updated docs, tests, and all references
