+++ plugins/sentry-cli/skills/sentry-cli/SKILL.md
@@ -323,10 +323,11 @@ Manage Sentry issues
 
 ### Event
 
-View and list Sentry events
+View, list, and send Sentry events
 
 - `sentry event view <org/project/event-id...>` — View details of one or more events
 - `sentry event list <issue>` — List events for an issue
+- `sentry event send <args...>` — Send a Sentry event
 
 → Full flags and examples: `references/event.md`
 
@@ -485,15 +486,6 @@ Browse the Sentry API schema
 - `sentry schema <resource...>` — Browse the Sentry API schema
 
 → Full flags and examples: `references/schema.md`
-
-### Send
-
-Send events and envelopes to Sentry via DSN
-
-- `sentry send event <args...>` — Send a Sentry event
-- `sentry send envelope <path...>` — Send a Sentry envelope file
-
-→ Full flags and examples: `references/send.md`
 
 ## Global Options
