+++ src/commands/event/index.ts
@@ -1,19 +1,22 @@
 import { buildRouteMap } from "../../lib/route-map.js";
 import { listCommand } from "./list.js";
+import { sendCommand } from "./send.js";
 import { viewCommand } from "./view.js";
 
 export const eventRoute = buildRouteMap({
   routes: {
     view: viewCommand,
     list: listCommand,
+    send: sendCommand,
   },
   defaultCommand: "view",
   docs: {
-    brief: "View and list Sentry events",
+    brief: "View, list, and send Sentry events",
     fullDescription:
-      "View and list event data from Sentry.\n\n" +
+      "View, list, and send event data from Sentry.\n\n" +
       "Use 'sentry event view <event-id>' to view a specific event.\n" +
-      "Use 'sentry event list <issue-id>' to list events for an issue.",
+      "Use 'sentry event list <issue-id>' to list events for an issue.\n" +
+      "Use 'sentry event send -m <message>' to send a test event.",
     hideRoute: {},
   },
 });
