+++ test/commands/event/send.test.ts
@@ -1,5 +1,5 @@
 /**
- * Tests for `sentry send-event` command func().
+ * Tests for `sentry event send` command func().
  */
 
 import {
@@ -11,10 +11,10 @@ import {
   spyOn,
   test,
 } from "bun:test";
-import { sendEventCommand } from "../../src/commands/send-event.js";
+import { sendCommand } from "../../../src/commands/event/send.js";
 // biome-ignore lint/performance/noNamespaceImport: needed for spyOn
-import * as transport from "../../src/lib/envelope/transport.js";
-import { useTestConfigDir } from "../helpers.js";
+import * as transport from "../../../src/lib/envelope/transport.js";
+import { useTestConfigDir } from "../../helpers.js";
 
 useTestConfigDir("send-event-");
 
@@ -37,12 +37,12 @@ function makeContext() {
   };
 }
 
-describe("sendEventCommand.func()", () => {
-  let func: Awaited<ReturnType<typeof sendEventCommand.loader>>;
+describe("sendCommand.func()", () => {
+  let func: Awaited<ReturnType<typeof sendCommand.loader>>;
   let sendSpy: ReturnType<typeof spyOn>;
 
   beforeEach(async () => {
-    func = await sendEventCommand.loader();
+    func = await sendCommand.loader();
     sendSpy = spyOn(transport, "sendEnvelopeRequest").mockResolvedValue(
       undefined
     );
@@ -126,7 +126,7 @@ describe("sendEventCommand.func()", () => {
 
   test("nonexistent file throws ValidationError (not raw stack trace)", async () => {
     const { ctx } = makeContext();
-    const { ValidationError } = await import("../../src/lib/errors.js");
+    const { ValidationError } = await import("../../../src/lib/errors.js");
     await expect(
       func.call(
         ctx,
@@ -138,7 +138,7 @@ describe("sendEventCommand.func()", () => {
 
   test("--raw requires file arguments", async () => {
     const { ctx } = makeContext();
-    const { ValidationError } = await import("../../src/lib/errors.js");
+    const { ValidationError } = await import("../../../src/lib/errors.js");
     await expect(
       func.call(ctx, { dsn: SAAS_DSN, raw: true, "no-environ": true })
     ).rejects.toBeInstanceOf(ValidationError);
